glowypa/Acne-detection-and-treatment-recommendations/backend/fastapi_all/requirements.txt
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
albucore
albumentations
annotated-types
anyio
appnope
asgiref
asttokens
attrs
backcall
backoff
bcrypt
beautifulsoup4
blis
numpy
build
cachetools
catalogue
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cloudpathlib
coloredlogs
comm
confection
contourpy
cryptography
cycler
cymem
dataclasses-json
debugpy
decorator
deepsearch-glm
Deprecated
dirtyjson
distro
dnspython
docling
docling-core
docling-ibm-models
docling-parse
docutils
docx2txt
durationpy
easyocr
elastic-transport
elasticsearch
email_validator
eval_type_backport
exceptiongroup
executing
fastapi
fastapi-cli
fasttext-wheel
filelock
filetype
fire
flatbuffers
fonttools
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
imageio
importlib-metadata
importlib_resources
ipykernel
ipython
jedi
Jinja2
jiter
joblib
jsonlines
jsonpatch
jsonpointer
jsonref
jsonschema
jsonschema-specifications
JsonWeb
jupyter_client
jupyter_core
kiwisolver
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langcodes
langdetect
langsmith
language_data
lazy_loader
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
lxml
marisa-trie
Markdown
markdown-it-py
marko
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mean-average-precision
mmh3
monotonic
motor
mpmath
multidict
murmurhash
mypy-extensions
nest-asyncio
networkx
ninja
nltk
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
ollama
onnxruntime
openai
opencv-python
opencv-python-headless
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
parso
pdfminer.six
pexpect
pickleshare
pillow
platformdirs
posthog
preshed
prompt-toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pybboxes
pybind11
pyclipper
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pymongo
PyMuPDF
pyparsing
pypdf
pypdfium2
PyPika
pyproject_hooks
python-bidi
python-dateutil
python-docx
python-dotenv
python-multipart
python-pptx
pytz
PyYAML
pyzmq
redis
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
Rtree
safetensors
sahi
scikit-image
scikit-learn
scipy
seaborn
sentence-transformers
setuptools
shapely
shellingham
six
smart-open
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
stack-data
starlette
striprtf
sympy
tabulate
tenacity
termcolor
terminaltables
thinc
threadpoolctl
tifffile
tiktoken
timm
tokenizers
tomli
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
triton
typer
typing-inspect
typing_extensions
tzdata
ultralytics
ultralytics-thop
uritemplate
urllib3
uvicorn
uvloop
wasabi
watchfiles
wcwidth
weasel
websocket
PyJWT